+++ packages/webview_flutter/webview_flutter_android/android/src/main/java/io/flutter/plugins/webviewflutter/WebViewClientFlutterApiImpl.java
@@ -8,6 +8,7 @@ import android.annotation.SuppressLint;
 import android.os.Build;
 import android.webkit.WebResourceError;
 import android.webkit.WebResourceRequest;
+import android.webkit.WebResourceResponse;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 import androidx.annotation.RequiresApi;
@@ -65,6 +66,16 @@ public class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
     return requestData.build();
   }
 
+  @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+  static GeneratedAndroidWebView.WebResourceResponseData createWebResourceResponseData(
+      WebResourceResponse response) {
+    final GeneratedAndroidWebView.WebResourceResponseData.Builder responseData =
+        new GeneratedAndroidWebView.WebResourceResponseData.Builder()
+            .setStatusCode((long) response.getStatusCode());
+
+    return responseData.build();
+  }
+
   /**
    * Creates a Flutter api that sends messages to Dart.
    *
@@ -97,6 +108,25 @@ public class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
     final Long webViewIdentifier =
         Objects.requireNonNull(instanceManager.getIdentifierForStrongReference(webView));
     onPageFinished(getIdentifierForClient(webViewClient), webViewIdentifier, urlArg, callback);
+  }
+
+  /** Passes arguments from {@link WebViewClient#onReceivedHttpError} to Dart. */
+  @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+  public void onReceivedHttpError(
+      WebViewClient webViewClient,
+      WebView webView,
+      WebResourceRequest request,
+      WebResourceResponse response,
+      Reply<Void> callback) {
+    webViewFlutterApi.create(webView, reply -> {});
+
+    final Long webViewIdentifier = instanceManager.getIdentifierForStrongReference(webView);
+    onReceivedHttpError(
+        getIdentifierForClient(webViewClient),
+        webViewIdentifier,
+        createWebResourceRequestData(request),
+        createWebResourceResponseData(response),
+        callback);
   }
 
   /**