+++ packages/webview_flutter/webview_flutter_android/android/src/main/java/io/flutter/plugins/webviewflutter/WebViewClientHostApiImpl.java
@@ -11,6 +11,7 @@ import android.os.Build;
 import android.view.KeyEvent;
 import android.webkit.WebResourceError;
 import android.webkit.WebResourceRequest;
+import android.webkit.WebResourceResponse;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 import androidx.annotation.NonNull;
@@ -52,6 +53,12 @@ public class WebViewClientHostApiImpl implements GeneratedAndroidWebView.WebView
     @Override
     public void onPageFinished(WebView view, String url) {
       flutterApi.onPageFinished(this, view, url, reply -> {});
+    }
+
+    @Override
+    public void onReceivedHttpError(
+        WebView view, WebResourceRequest request, WebResourceResponse response) {
+      flutterApi.onReceivedHttpError(this, view, request, response, reply -> {});
     }
 
     @Override
@@ -115,6 +122,13 @@ public class WebViewClientHostApiImpl implements GeneratedAndroidWebView.WebView
     @Override
     public void onPageFinished(WebView view, String url) {
       flutterApi.onPageFinished(this, view, url, reply -> {});
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    @Override
+    public void onReceivedHttpError(
+        @NonNull WebView view, @NonNull WebResourceRequest request, WebResourceResponse response) {
+      flutterApi.onReceivedHttpError(this, view, request, response, reply -> {});
     }
 
     // This method is only called when the WebViewFeature.RECEIVE_WEB_RESOURCE_ERROR feature is