+++ packages/webview_flutter/webview_flutter_wkwebview/lib/src/web_kit/web_kit_api_impls.dart
@@ -73,6 +73,16 @@ extension _WKNavigationActionPolicyConverter on WKNavigationActionPolicy {
   }
 }
 
+extension _WKNavigationResponsePolicyConverter on WKNavigationResponsePolicy {
+  WKNavigationResponsePolicyEnumData toWKNavigationResponsePolicyEnumData() {
+    return WKNavigationResponsePolicyEnumData(
+      value: WKNavigationResponsePolicyEnum.values.firstWhere(
+        (WKNavigationResponsePolicyEnum element) => element.name == name,
+      ),
+    );
+  }
+}
+
 extension _NSHttpCookiePropertyKeyConverter on NSHttpCookiePropertyKey {
   NSHttpCookiePropertyKeyEnumData toNSHttpCookiePropertyKeyEnumData() {
     late final NSHttpCookiePropertyKeyEnum value;
@@ -176,6 +186,13 @@ extension _NavigationActionDataConverter on WKNavigationActionData {
   }
 }
 
+extension _NavigationResponseDataConverter on WKNavigationResponseData {
+  WKNavigationResponse toNavigationResponse() {
+    return WKNavigationResponse(
+        response: response.toNSUrlResponse(), forMainFrame: forMainFrame);
+  }
+}
+
 extension _WKFrameInfoDataConverter on WKFrameInfoData {
   WKFrameInfo toWKFrameInfo() {
     return WKFrameInfo(isMainFrame: isMainFrame);
@@ -190,6 +207,12 @@ extension _NSUrlRequestDataConverter on NSUrlRequestData {
       httpMethod: httpMethod,
       allHttpHeaderFields: allHttpHeaderFields.cast(),
     );
+  }
+}
+
+extension _NSUrlResponseDataConverter on NSHttpUrlResponseData {
+  NSHttpUrlResponse toNSUrlResponse() {
+    return NSHttpUrlResponse(statusCode: statusCode);
   }
 }
 
@@ -842,6 +865,31 @@ class WKNavigationDelegateFlutterApiImpl
           as WKWebView,
       url,
     );
+  }
+
+  @override
+  Future<WKNavigationResponsePolicyEnumData> decidePolicyForNavigationResponse(
+    int identifier,
+    int webViewIdentifier,
+    WKNavigationResponseData navigationResponse,
+  ) async {
+    final Future<WKNavigationResponsePolicy> Function(
+      WKWebView,
+      WKNavigationResponse navigationResponse,
+    )? function = _getDelegate(identifier).decidePolicyForNavigationResponse;
+
+    if (function == null) {
+      return WKNavigationResponsePolicyEnumData(
+        value: WKNavigationResponsePolicyEnum.allow,
+      );
+    }
+
+    final WKNavigationResponsePolicy policy = await function(
+      instanceManager.getInstanceWithWeakReference(webViewIdentifier)!
+          as WKWebView,
+      navigationResponse.toNavigationResponse(),
+    );
+    return policy.toWKNavigationResponsePolicyEnumData();
   }
 
   @override