+++ packages/webview_flutter/webview_flutter_wkwebview/lib/src/webkit_webview_controller.dart
@@ -615,6 +615,16 @@ class WebKitNavigationDelegate extends PlatformNavigationDelegate {
           weakThis.target!._onPageStarted!(url ?? '');
         }
       },
+      decidePolicyForNavigationResponse:
+          (WKWebView webView, WKNavigationResponse response) async {
+        if (weakThis.target?._onHttpError != null &&
+            response.response.statusCode >= 400) {
+          weakThis.target!._onHttpError!(
+              HttpResponseError(statusCode: response.response.statusCode));
+        }
+
+        return WKNavigationResponsePolicy.allow;
+      },
       decidePolicyForNavigationAction: (
         WKWebView webView,
         WKNavigationAction action,
@@ -688,6 +698,7 @@ class WebKitNavigationDelegate extends PlatformNavigationDelegate {
 
   PageEventCallback? _onPageFinished;
   PageEventCallback? _onPageStarted;
+  HttpResponseErrorCallback? _onHttpError;
   ProgressCallback? _onProgress;
   WebResourceErrorCallback? _onWebResourceError;
   NavigationRequestCallback? _onNavigationRequest;
@@ -700,6 +711,11 @@ class WebKitNavigationDelegate extends PlatformNavigationDelegate {
   @override
   Future<void> setOnPageStarted(PageEventCallback onPageStarted) async {
     _onPageStarted = onPageStarted;
+  }
+
+  @override
+  Future<void> setOnHttpError(HttpResponseErrorCallback onHttpError) async {
+    _onHttpError = onHttpError;
   }
 
   @override